Organize test files: Move all tests and related files to tests directory
- Moved all test_*.py files to tests/
- Moved all JSON test result files to tests/
- Moved debug_*.py scripts to tests/
- Moved analyze_*.py scripts to tests/
- Moved utility scripts (create_test*, check_*) to tests/
- Moved test data files (*.json, *.xlsx, *.html) to tests/
- Updated tests/README.md with comprehensive documentation of all test files
- Organized documentation into logical sections: Core Tests, PDF Processing, Web Interface, Debug Scripts, Utilities
- Added detailed descriptions for each test file and its purpose
