+++ tests/check_frozen_panes.py
