+++ tests/debug_table_detection.py
