+++ tests/debug_table_regions.py
