+++ tests/test_decimal_with_suffix.py
