+++ tests/test_financial_table_detection.py
