+++ tests/test_table_100_numbers.py
