+++ tests/test_web_interface_display.py
